+++ src/object/plan.c
@@ -6,7 +6,7 @@
 /*   By: glodenos <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/22 21:51:55 by glodenos          #+#    #+#             */
-/*   Updated: 2016/10/04 22:29:59 by glodenos         ###   ########.fr       */
+/*   Updated: 2016/10/05 15:27:04 by glodenos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ inline void plan(register t_obj *obj, register t_ray ray)
 
     tmp_v = vector_sub(obj->pos, origine);
     normal = vector_normalize(obj->rotate);
-    d = sqrt(vector_scalar(tmp_v, tmp_v));
+    d = sqrtf(vector_scalar(tmp_v, tmp_v));
     det = -(normal.x * ray.a.x + normal.y * ray.a.y + normal.z * ray.a.z + d)
         / (normal.x * ray.b.x + normal.y * ray.b.y + normal.z * ray.b.z);
     obj->det = (det < 0) ? -1 : det;
