+++ src/vector/coordinates_collision.c
@@ -6,14 +6,14 @@
 /*   By: glodenos <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/23 03:56:23 by glodenos          #+#    #+#             */
-/*   Updated: 2016/10/04 22:26:21 by glodenos         ###   ########.fr       */
+/*   Updated: 2016/10/05 16:12:41 by glodenos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "lib_RT.h"
 
 inline cl_float3    coordinates_collision(register cl_float3 a,
-        register cl_float3 b, register double det)
+        register cl_float3 b, register float det)
 {
     register cl_float3  vect;
